migrate code co modular scripts
