+++ models/purple_alien/main.py
@@ -1,13 +1,13 @@
-import numpy as np
-import pickle
+#import numpy as np
+#import pickle
 import time
-import os
-import functools
-from datetime import datetime
+#import os
+#import functools
+#from datetime import datetime
 
-import torch
-import torch.nn as nn
-import torch.nn.functional as F
+#import torch
+#import torch.nn as nn
+#import torch.nn.functional as F
 
 import wandb
 
@@ -19,17 +19,19 @@ sys.path.insert(0, str(Path(*[i for i in PATH.parts[:PATH.parts.index("views_pip
 from set_path import setup_project_paths, setup_artifacts_paths
 setup_project_paths(PATH)
 
-from utils import choose_model, choose_loss, choose_sheduler, get_train_tensors, get_full_tensor, apply_dropout, execute_freeze_h_option, get_log_dict, train_log, init_weights, get_data
-from utils_wandb import add_wandb_monthly_metrics
-from utils_device import setup_device
-from config_sweep import get_swep_config
-from config_hyperparameters import get_hp_config
-from train_model import make, training_loop, handle_training
+#from utils import choose_model, choose_loss, choose_sheduler, get_train_tensors, get_full_tensor, apply_dropout, execute_freeze_h_option, get_log_dict, train_log, init_weights, get_data
+#from utils_wandb import add_wandb_monthly_metrics
+#from utils_device import setup_device
+#from config_sweep import get_swep_config
+#from config_hyperparameters import get_hp_config
+#from train_model import make, training_loop, handle_training
 # from evaluate_sweep import evaluate_posterior # see if it can be more genrel to a single model as well... 
-from evaluate_model import evaluate_posterior
+#from evaluate_model import evaluate_posterior, handle_evaluation
+#from forecast_model import handle_forecasting
 from cli_parser_utils import parse_args, validate_arguments
-from artifacts_utils import get_latest_model_artifact
+#from artifacts_utils import get_latest_model_artifact
 
+#from mode_run_manager import model_run_manager
 
 #def setup_device(): 
 #    # set the device
@@ -43,42 +45,42 @@ from artifacts_utils import get_latest_model_artifact
 #    wandb.define_metric("monthly/*", step_metric="monthly/out_sample_month")
                 
 
-def handle_sweep_run(args):
-    print('Running sweep...')
-
-    project = f"purple_alien_sweep" # check naming convention
-    sweep_config = get_swep_config()
-    sweep_config['parameters']['run_type'] = {'value' : "calibration"} # I see no reason to run the other types in the sweep
-    sweep_config['parameters']['sweep'] = {'value' : True}
-
-    sweep_id = wandb.sweep(sweep_config, project=project) # and then you put in the right project name
-
-    wandb.agent(sweep_id, model_pipeline)
-
-
-def handle_single_run(args):
-    
-    # get hyperparameters. IS THE ISSUE UP HERE?
-    hyperparameters = get_hp_config()
-    hyperparameters['run_type'] = args.run_type
-    hyperparameters['sweep'] = False
-
-    # get run type and denoting project name - check convention!
-    project = f"purple_alien_{args.run_type}"
-
-    if args.run_type == 'calibration' or args.run_type == 'testing':
-
-        model_pipeline(config = hyperparameters, project = project, train = args.train, eval = args.evaluate, forecast = False, artifact_name = args.artifact_name)        
-
-    elif args.run_type == 'forecasting':
-
-        #print('True forecasting ->->->->')
-        model_pipeline(config = hyperparameters, project = project, train = False, eval = False, forecast=True, artifact_name = args.artifact_name)     
-
-    else:
-        raise ValueError(f"Invalid run type: {args.run_type}")
-
-
+# def handle_sweep_run(args):
+#     print('Running sweep...')
+# 
+#     project = f"purple_alien_sweep" # check naming convention
+#     sweep_config = get_swep_config()
+#     sweep_config['parameters']['run_type'] = {'value' : "calibration"} # I see no reason to run the other types in the sweep
+#     sweep_config['parameters']['sweep'] = {'value' : True}
+# 
+#     sweep_id = wandb.sweep(sweep_config, project=project) # and then you put in the right project name
+# 
+#     wandb.agent(sweep_id, model_run_manager)
+# 
+# 
+# def handle_single_run(args):
+#     
+#     # get hyperparameters. IS THE ISSUE UP HERE?
+#     hyperparameters = get_hp_config()
+#     hyperparameters['run_type'] = args.run_type
+#     hyperparameters['sweep'] = False
+# 
+#     # get run type and denoting project name - check convention!
+#     project = f"purple_alien_{args.run_type}"
+# 
+#     if args.run_type == 'calibration' or args.run_type == 'testing':
+# 
+#         model_run_manager(config = hyperparameters, project = project, train = args.train, eval = args.evaluate, forecast = False, artifact_name = args.artifact_name)        
+# 
+#     elif args.run_type == 'forecasting':
+# 
+#         #print('True forecasting ->->->->')
+#         model_run_manager(config = hyperparameters, project = project, train = False, eval = False, forecast=True, artifact_name = args.artifact_name)     
+# 
+#     else:
+#         raise ValueError(f"Invalid run type: {args.run_type}")
+# 
+# 
 #def handle_training(config, device, views_vol, PATH_ARTIFACTS):
 #    
 #    # Create the model, criterion, optimizer and scheduler
@@ -150,49 +152,48 @@ def handle_single_run(args):
 #def handle_forecasting(config, device, views_vol, PATH_ARTIFACTS, artifact_name=None):
 #
 #    raise NotImplementedError('Forecasting not implemented yet')
+
 #
-
-
-def model_pipeline(config = None, project = None, train = None, eval = None, forecast = None, artifact_name = None):
-
-    # Define the path for the artifacts
-    PATH_ARTIFACTS = setup_artifacts_paths(PATH)
-
-    device = setup_device()
-
-    # Initialize WandB
-    with wandb.init(project=project, entity="views_pipeline", config=config): # project and config ignored when running a sweep
-        
-        # add the monthly metrics to WandB
-        add_wandb_monthly_metrics() 
-
-        # Update config from WandB initialization above
-        config = wandb.config
-
-        # Retrieve data (partition) based on the configuration
-        views_vol = get_data(config)
-
-        # Handle the sweep runs
-        if config.sweep:  # If we are running a sweep, always train and evaluate
-
-            model, criterion, optimizer, scheduler = make(config, device)
-            training_loop(config, model, criterion, optimizer, scheduler, views_vol, device)
-            print('Done training')
-
-            evaluate_posterior(model, views_vol, config, device)
-            print('Done testing')
-
-        # Handle the single model runs: train and save the model as an artifact
-        if train:
-            handle_training(config, device, views_vol, PATH_ARTIFACTS)
-
-        # Handle the single model runs: evaluate a trained model (artifact)
-        if eval:
-            handle_evaluation(config, device, views_vol, PATH_ARTIFACTS, artifact_name)
-
-        if forecast:
-            handle_forecasting(config, device, views_vol, PATH_ARTIFACTS, artifact_name)
-
+#def model_run_manager(config = None, project = None, train = None, eval = None, forecast = None, artifact_name = None):
+#
+#    # Define the path for the artifacts
+#    PATH_ARTIFACTS = setup_artifacts_paths(PATH)
+#
+#    device = setup_device()
+#
+#    # Initialize WandB
+#    with wandb.init(project=project, entity="views_pipeline", config=config): # project and config ignored when running a sweep
+#        
+#        # add the monthly metrics to WandB
+#        add_wandb_monthly_metrics() 
+#
+#        # Update config from WandB initialization above
+#        config = wandb.config
+#
+#        # Retrieve data (partition) based on the configuration
+#        views_vol = get_data(config)
+#
+#        # Handle the sweep runs
+#        if config.sweep:  # If we are running a sweep, always train and evaluate
+#
+#            model, criterion, optimizer, scheduler = make(config, device)
+#            training_loop(config, model, criterion, optimizer, scheduler, views_vol, device)
+#            print('Done training')
+#
+#            evaluate_posterior(model, views_vol, config, device)
+#            print('Done testing')
+#
+#        # Handle the single model runs: train and save the model as an artifact
+#        if train:
+#            handle_training(config, device, views_vol, PATH_ARTIFACTS)
+#
+#        # Handle the single model runs: evaluate a trained model (artifact)
+#        if eval:
+#            handle_evaluation(config, device, views_vol, PATH_ARTIFACTS, artifact_name)
+#
+#        if forecast:
+#            handle_forecasting(config, device, views_vol, PATH_ARTIFACTS, artifact_name)
+#
 
 if __name__ == "__main__":
 
@@ -212,7 +213,6 @@ if __name__ == "__main__":
     if args.sweep == True:
 
         handle_sweep_run(args)
- 
 
     elif args.sweep == False:
