+++ models/purple_alien/src/utils/utils_device.py
@@ -1,0 +1,7 @@
+import torch
+
+def setup_device():
+    # Set the device
+    device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
+    print(f"Using device: {device}")
+    return device   # not sure you need to return it, but it might be useful for debugging
